git_show_suse.sh: Leave top level directory only with option -x
When multiple top directories are found, we did not strip the
top level directory when using a rpm-git generated repository
to accomodate the oddities of Mesa. This is not really needed
any more. Do this only when option -x is specified.
Unless a commit has multiple different top directories: in this
case do not strip these either.

Signed-off-by: Egbert Eich <eich@suse.com>