Pass local RPM environment in RPMARGS to interactive subshell
Inside the interactive shell one can make use of these settings
when calling rpm(|build|spec):
IFS=@
rpm[|build|spec] ${RPMARGS} ...
(The individual arguments for rpm are separated by `@` to compensate
for a shortcoming of shell).
For this to work, `su` can no longer be used to change the user to the
target user. `chroot`'s `chroot --userspec=` comes to the rescue.

Signed-off-by: Egbert Eich <eich@suse.com>